requirements.txt
mysql-connector-python>=8.0.31
simple-term-menu>=1.5.3